Move DynawoCurveGroovyExtension into th dynawo package
