+++ src/main/groovy/com/powsybl/dynawo/DynawoCurveGroovyExtension.groovy
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.dynawo.dynamicmodels
+package com.powsybl.dynawo
 
 import com.google.auto.service.AutoService
 import com.powsybl.dsl.DslException
